pyproject.toml
[tool.poetry]
name = "fast-api-udemy2"
version = "0.1.0"
description = ""
authors = ["rjribeiro"]
#readme = "README.md"
packages = [{include = "./app"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.100.0"
pydantic = "^2.0.2"
uvicorn = "^0.22.0"
pytest = "^7.4.0"
sqlalchemy = "^2.0.18"
psycopg2-binary = "^2.9.6"
alembic = "^1.11.1"
python-decouple = "^3.8"
httpx = "^0.24.1"
flake8 = "^6.1.0"
passlib = "^1.7.4"
python-jose = "^3.3.0"
python-multipart = "^0.0.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
